Enhance: add better support for environments
Whilst checking my improvements to Map XML importing I noticed that the map
for Achaea had multiple different html colour names assigned to the same
1-15 range that we have been using.  This prompted a revision in the way
environments are processed...

If an (IRE MUDs in particular) XML map is imported we now parse the
"htmlcolor" and "name" tags in the environment data and use the former to
directly set the corresponding TMap::customEnvColor entry to that color.
This differed from past code that read just the "color" numeric value tag
and set a cross mapping from envId to colorId in what was TMap::envColors
but which has been renamed to mEnvColorIdMap. This map was INACCESSIBLE
to the Mudlet user but reserved the colour Ids of 1-16.  The code in this
commit now ignores the "color" tag, if it can, so that, particularly the
Achaea MUD can specify different actual colours for different Environment
Ids even though they specify the same color Id in the "color" tag.

Instead of picking out specific attributes for the environment / area /
room XML elements when parsing we now iterator through ALL attributes and
then store unknown / unexpected attributes within the map / area / room
user data for that item.  The key begins with "xmlParse", which should be
specifically reserved for this.  The constructed key uses the attribute's
name as found in the file so for the base room it is of the form:
"xmlParse.<atttibuteName>"
for a parameter from the coordinate sub-element the key is of the form:
"xmlParse.coord.<attributeName>"
and from an exit sub-element the key is of the form:
"xmlParse.exit.<exitname>.<attributeName>"
Similarly area related extra items are stored as:
"xmlParse.<atttibuteName>" in the relevant area User Data structure
and extra environment items in the map User Data structure as:
"xmlParse.environment.<env Id number>.<atttibuteName>"

Special exits are now parsed.

A room or exit attribute {not used by IRE MUD maps} "locked" equal to
"yes", "1" or "true" will set the corresponding speed walk "don't use this"
room/exit flag.

Similarly a room or exit attribute {not used by IRE MUD maps} "cost" being
the text equivalent of a positive integer will set the corresponding
exit or room weight value.

The raw "door" and "hidden" attributes for exits are written into the room
user data as well as being combined and used to set the room exit's door
as appropriate (this is so that if the latter is not right the user can
tell what the original data elements were.)

As a consequence - the use of "building" for IRE MUDs maps as a room
coordinate sub-element attribute can now be seen by the user as requested
in forum post: http://forums.mudlet.org/viewtopic.php?f=13&t=4914 !

The following lua commands are added:
* getEnvironmentTable() - which returns the entire stored environment data
  in a table with envID as keys and values of:
  { ["name"] = "<text assigned to environment>",
    ["colorId"] = colorId,
    ["defaultColor"] = { r, g, b, a } }
  colorId is either an index to one of the 16 mapper colours, which can be
  adjusted by the mapper colours on the profile preferences and for all but
  IRE MUDs apart from Achaea {Aetolia, Imperian, Lusternia &
  Midkemia-Online} seem to be a fixed set at the present; or one of the
  customEnvColors - which will be a "straight-through" mapping using the
  SAME value as the key.
  defaultColor is a four colour component table WHICH IS OMITTED currently
  (both key and value) if no htmlcolor tag was parsed for this environment.

* getEnvironment((number)id) - returns the string associated with the
  given environment Id number.

In addition, provision is made for:
* setEnvironment((number)id, (string)name[, (number) redComponent,
                 (number) greenComponent, (number) blueComponent
                 [, (number) alphaComponent]])
* resetEnvironment((number)id)
  which allows the user to add and removes texts for the given environment
  id, for the setter an optional 3 or 4 colour component arguments permit
  direct setting of the corresponding customEnvColor if the id is greater
  than 16.  They both manipulate TMap::mEnvColorIdMap and also the new
  TMap::mEnvColorNamesMap and TMap::mEnvColorNamesMap

However, although the code is present for these functions THEY NEED the map
file format to be increased to 17 because storage of TMap::mEnvColorNamesMap
and TMap::mEnvColorNamesMap is required needing revision to that.
They will emit a warning to the console if the map file format for saving is
not at the required setting incremented - which, following the use of the
manual override on the "Special" options (last) tab of the Profile Preference
dialog can be done.

Revision has been done to both 2D and 3D mappers to extend them to use the
new environment/colour specification systems.

Also:
* Comment out a pair SIGNAL-SLOT connections in dlgProfilePreferences()
  that fail to connect to non-existent slots.
* Bring up to current coding style TLuaInterpreter::getRoomEnv().
* Remove unused global QMap<int,int> areaMap from XMLimport class.
* In anticipation of possibly adding a GUI interface for environments to
  the profile_preference dialogue I have sorted the many grid_layout parts
  to be in row/column order so that a future diff of any change will
  produce less noise - there should not be any functional change from this
  commit.
* Fixed a one-line error in GLWidget::paintGL() where an exit was marked
  as being an area exit when it isn't.

Re-based and resolved conflicts following the movement of the code that
created the map and area user data structure in the map to a separate,
now preceding, branch/Pull Request in:
	src/ui/profile_preferences.ui
	src/TLuaInterpreter.cpp
	src/TLuaInterpreter.h
	src/TMap.cpp
	src/TMap.h

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>