Revise: update TLuaInterpreter::loadMap() to handle XML files
If the optional file to load ends with ".xml" the revised loadMap() command
will use a TConsole::importMap() call instead of TConsole::loadMap() so
that such files can be read.  This could be of benefit for those who have
not been able to successfully use the existing automated download and load
capabilities or who want to script an importation process from another
MUD client.

Whilst updating the coding and message style of this command to current
characteristics I also brought the corresponding saveMap command to the
same standards - though as we do not currently have XML map export
capabilities (yet?) trying to save a map with an ".xml" extension will
produce a nil and warning message return values.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>