Revision: change Travis CI to build C & Q make projects in separate builds
This restores building both the Linux qmake and cmake project files but
splits them up into two separate builds.  As cmake is present for both
platforms I think we should also do this for MacOs which this commit should
now do.  Given that the primary use case for CMake appears (to me) to be
the (as yet not supported by Travis) Windows platform I have marked the
cmake tests to be allowed to fail as it is not obvious that the cmake
project files are fully set up for either of the current platforms.

The repository version of this branch has slipped behind it's main-line
parent (release_30) and I have rebased my local working copy to a later
point in that - as a result the repository branch requires a FORCED PUSH
and so the history has been rewritten (as I understand it)...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>