FixEnhance: Improve Travis C.I. testing process.
Split the Linux build into two separate process using the qmake and cmake
project files.  Whilst the previous code for the qmake file worked with
the MacOs platform it was not possible to build Mudlet with the cmake
project files.  This commit tweaks various things so that it now does.

For Macintosh computers there is not an easy way to obtain the lua zip
module as a pre-compiled package so we have to include the source code
ourselves.  In checking this I found that we did not have all the text
that I think we needed to satisfy the terms of the Licence (which is NOT
GPL, though stated to be compatible with it).

Added the CMake project and some related files to the qmake project file
so that they can be found and edited whilst working in QtCreator from that
file.

Also put quotation marks around build designator in Top Level CMake project
file so that its structure matches the one in the other branch.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>