FixEnhance: Improve Travis C.I. testing process.
This commit is a rebase of a whole series that I put out to achieve
everything it now does. This message should represent a reasonable
summary of the end results.

Split the Linux build into two separate process using the qmake and cmake
project files.  Whilst the previous code for the qmake file worked with
the MacOs platform it was not possible to build Mudlet with the cmake
project files.  This commit tweaks various things so that it now does.

For Macintosh computers there is not an easy way to obtain the lua zip
module as a pre-compiled package so we have to include the source code
ourselves.  In checking this I found that we did not have all the text
that I think we needed to satisfy the terms of the Licence (which is NOT
GPL, though stated to be compatible with it).

Added the CMake project and some related files to the qmake project file
so that they can be found and edited whilst working in QtCreator from that
file.

Clean up a couple of things spotted during C.I. compilations:

* In TRoomDB::__removeRoom(int id) there was an ambiguous logic construct
which needed some parenthesis to clarify the intent.

* In TRoomDB::setAreaName(int areaID, QString name) an erronous printf
format specifier was included in a debug warning message.

* The CMake build version variable was out of date and did not match the
QMake one - given that CMake is the likely build method for Windows it
suggests that there might be preview version in the field that declare
themselves to be "3.0.0-beta" when they are actually "-gamma" or "-delta"
versions!

Include some Travis files into the project file so that they show up in the
Qt Creator IDE (as they now do in the other branches version).

Copy across the QMAKE_CXXFLAGS and other tweaks from the other branch
version. This includes the C++11 enabling Qt switch.

MacOs C++11 support seems to need MacOs >10.6 though I could easily be wrong!

Change Travis setup to use both CLang & GCC: since we are testing
compilers we can try doubling the build matrix to use both Clang and GCC
compilers...

I've removed qt4-dev tools from the Linux lineup as we no longer support
it and it is more tricky to work with both Qt4 and Qt5 development tools
in place.

The GCC update for Linux was ported from Development branch of code.

Remove old boost packages from Linux Travis Builds as I found that the
explicitly numbered libboost1.46 package was not being replaced by
unnumbered upgrade (1.48) and I think that was causing compilation
problems for the Linux qmake entries in the CI matrix.

Put in a command to purge the unused auto-dependences that the qt4
removed items orphans - which otherwise meant time was being spent
upgrading pointlessly the unremoved ones.

As it seems there is a problem in the Qt 5.0.2 libraries that the Linux
clang compiler objects to but the gcc 4.7 does not when run via cmake I
have put in an "allow failure" for that row of the build matrix.

Also ensure boost 1.48 is installed uniformly, the default seems to mix
up 1.46 and 1.48 components.

The test platform only has Qt5.0.2 for Linux and this, used with CMake,
seems to fail internally within the Qt header files.

Also added a couple of lines to report the compiler to be used AFTER it
may have been updated during the installation stages (the system does
report it but AT THE BEGINNING before package updates are carried out).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>