Fixup: remove qmake build from linux C.I. and add .travis file to project
The linux build does a qmake build before it does a cmake one - this sames
to be unnecessary.

A warning message from the MacOs C.I. builds advises that the Qt5
installation on the Travis host is a "keg" build and that the command
"brew linkapps qt" is advisable and/or including in the environment:
LDFLAGS += -L/usr/local/opt/qt5/lib
CPPFLAGS += -I/usr/local/opt/qt5/include

The CI control file .travis was missing from the Qt project file src.pro.
Also there was a couple of tweaks needed there for the QMAKE_CXX_FLAGS:
the release one needed to have any preset optimisatin flags removed before
we set our own (-03) for the RELEASE one and we do not need to put in -g
for the DEBUG one (QT puts THAT in automatically).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>