Fix bug resulting in duplicate .xml when exporting profiles or triggers
The functions responsible for exporting profiles and triggers were
blindly appending ".xml" onto the end of the filename before writing.
This would result in exporting a file called <name>.xml.xml if the
user specified <name>.xml. This commit adds a case-insensitive check
to see if the filename already ends in ".xml" before appending it
to the end of the filename.