BugFix: Several one-liner fixes
Two broken connections in dlgProfilePreferences c'tor to non-existent
dlgProfilePreferences::setBorderColor() and ::setBorderImage() slots from
a pair of QPushButtons on the Profile Preferences dialogue.  Commented out
connections to suppress run-time error message when the dialogue is
invoked.  There is a TLuaInterpreter::setBorderColor() but this directly
sets the colour of three parts of the palette of the frame of the main
profile console window.

Commented out two unneeded qDebug() messages related to the left and right
boarder widths stored in (Host *)->mBorderLeftWidth and
->mBorderRightWidth .

Fixed a typo in dlgTriggerEditor::slot_script_toggle_active() that caused
an icon to be missing when a script is activated by saving.

Fixed a missing "else" in dlgTriggerEditor::saveAlias() that caused an
icon to always be the grey checked one whether the ancestors were active
(when the icon should be the normal coloured one) or not (when the icon
should be grey).

Text for top line of "Keys" treeWidget was being sent to wrong variable so
was not showing.

Icon assigned to main application and thus "Alt-Tab" icon was the lowest
resolution icon available only 16x16 and users had noticed, changed to
48x48 one which is not too much better for high resolution display users
(Mac "Retina" displays?) but is significantly improved for others.

Add existing but unused icon to profile notes window (and configure the
title to be translatable).

Add newly created mudlet_irc.png icon to IRC window (and configure the
title to be translatable).  The icon was created by me from the Mudlet
48x48 pixel image with the Oxygen/KDE team's internet-telephony.png 22x22
pixel image superimposed in bottom right quarter.  Both source images are
believed to be GPLv2+ or compatible and so I licence this image under
GPLv2+ terms also.

In several places the constant QFont::Courier is used when it is not a
valid value to use - as a Font Weight argument in a QFont constructor.
This enum constant (seems to be integer value 8) is a QFont::StyleHint
whereas the requirement is for a QFont::Weight (range of weights is 0 to
99) and typical predefined values include QFont::Normal (50) and
QFont::Bold (75) given that most fonts may only have one or two weights
defined and typically these may be of these pair anyway, under most
situations where QFont::Courier is mistakenly used the nearest available
value will be the QFont::Normal weight - which is the case with our default
of the Bitstream Vera Sans Mono font we use in many places.  So it would
seem wise to put in the value that is available rather than rely on the
font system fixing things magically.

Ensured status bar is created at bottom of application window, getting a
pointer to it causes it to be generated - and some parts of the application
do use it!

Fixed a coding error where areaExit was set to true for both area and
non-area exits in glwidget::paintGL() - this cause bogus rooms to appear
on rooms on Z-levels about the level of the current view centre room!

Changed return type and renamed (int)Host::check_for_mappingscript() to
(const bool)Host::checkForMappingScript() to reflect that it does return
a boolean.

BugFix for replay recording code reporting wrong logfile name when the
recording is stopped (it was returning the plain text/HTML log file name
NOT the replay one!)

BugFix for intermittent replay playback not decoding colour information in
replay, caused by TBuffer::mIsHighColorMode not being initialised before
use!

BugFixes for:
A) invisible replay speed control buttons (icons size was not specified);
B) no time display (zero time is invalid and since Qt 5.0 adding anything
to it is STILL invalid).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>