Enhance: Add control for display of White-space characters in Editor window
Whilst perusing the Qt Documentation I found the options to display Tabs &
Space (as right point arrows and centre dots) and Line-Feeds & Paragraphs
(as down then left arrow and pilcrow) respectively. As this might be useful
in some circumstance I have added user controllable options for these two
settings in the "Main Display" tab of the "Profile Preferences" dialog.
The setting is applied to any and all "editor" areas (if multiple profiles
are active) and is maintained/persists in the main config file
(~/.config/Mudlet/Mudlet 1.0.conf).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>