Tidy: Further tidy of trigger_editor form
This commit removes several nested QWidgets that add nothing to the overall
designed but were probably an artifcate of the evolution of this form.

As this change will create a massive (and "noisy") difference to the file
it has been done separately so that the previous commit changes do not
get "lost" amongst the effects of this one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>