Enhance: tweaks to Source Editor area
Make background area of source editor different from that of content so
the end of the later can be distinguished.

Add placeholder text with a couple of hints when the source editor area is
empty - as this was not imoplimented for the widget used until Qt 5.2 and
we need to compile for  versions prior to that wrap the relevant part in
QT_VERSION conditional compilation directives.

Fix the wrong line count details (shows total line count of 2 for an empty
widget) - in doing this found it was possible to include character count
details both for whole widget contents and for the current block (a.k.a.
line/paragraph).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>