Enhance: allow Main window statusBar permanently shown/auto-hidable/hidden
Having recent uncovered and reactivated the status bar that was coded for
but not showing it was found that some users may not appreciated the space
that it could take up on a small screen platform.  With a bit of code
I worked out how to make it "pop-up" only when needed and have added that
functionality - as well as a control to set the mode for the status bar,
being one of: "Off", "Auto" or "On".

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>