Fixup: provide statusBarTips for main menu actions
Tooltips on QMenu items are NOT displayed by QT - a known (mis)feature?
The recommended action is to use statusBarTips instead.  This commit
provides them for the main menu item.

Note: THE MAIN MENU DOES NEED OVERHAULING:#
* some items are not disabled when they should be in the manner that the
corresponding toolbar buttons are
* access to the editor is limited to the "scripts" entry point, unlike
the individual toolbar button
* the invitation to "download the latest version" is laughably wrong in
that it sends the user to the old SOURCEFORGE repository.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>