Bugfix: correct errors in Lua command error messages
There has been a fair amount of copy and paste in creating various Lua
commands - but the error messages have not always been revised when this
has happened.  With a bit of *nix shell-fu I have identified the cases
where this has happened and changed the function that is given as the
name where an error occurred.

Note that this does not attempt to improve the formatting to bring the
standard up to current guidelines - or that the amended functions actually
operate as intended - merely it fixes gross errors that will definitely be
confusing to anyone seeing those messages.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>