Enhance: make map auditing more thorough and robust
As well as fixing the Area ownership of rooms (which has been an issue that
crops up from past buggy TArea code) this also clears out custom exit line
junk - including details for lines where the key was an empty string, those
arose during previous buggy code that allow exit lines to be drawn without
specifying which exit they were for.

The checks also removes any stub exits that are present on rooms that have
an actual exit in the same (normal exit) direction.  The current code
prevents the creation of such things but it did not in the past (I have
encountered it - especially when the real exit is a "circular" one that
goes to the SAME room as it leaves from!)  BTW Such a exit type, although
initially seeming to be pointless is a feature of some MUDs and the normal
exit drawing code for the 2D map does not currently show them (future
thing to cover?) though a custom exit line can be drawn that shows the
"circular-ness" of the exit!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>