Enhance: add ability to get ALL keys used by ALL room/area User Data items
Not essential but can be useful when working with maps not of own making
as it enable the user to quickly find everything that they can use with
other functions rather than having to iterate through each room/area in the
entire map.

Also remove a duplicate entry for searchRoomUserData that slipped into
TLuaInterpreter::initLuaGlobals().

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>