Fixup: make copying map to multiple profiles work better
Refactor the code that identifies where player is in other profiles, saves
it in current profile's map, copies it to other profiles and tells them
to load it in.

The code to copy to another map as was only permitted one other profile to
be done at a time - this makes it difficult to ensure that all the profiles
get the same data for all the other affected maps. The profile selection
part of "Profile Preferences" "Mapper" has been changed to be something
that allows all the destination targets (other profiles) to be selected,
excluding the current profile (as that ALWAYS gets the new map - it is
intrinsic to the copying process).  Only when one (or more) other profiles
are selected does the "Copy" button become enabled and then all active
profiles are checked to see if they are included, their player room is
noted as before and all the details are written to the current profile's
new map.  After copying (with some checks for success) the map to ALL
profiles a signal is sent to all active profiles (with a list of profiles
to respond) that causes the listed ones to reload a new map from their
home map directory.  (Extensive, formatted!) tooltips have been added to
the two controls concerned in the profile preferences.

Slots and signals are used (via the singleton class mudlet) to try and
accommodate future situations when Mudlet becomes properly multi-threaded
and each profile runs in a different thread - getting the current room
from each profile may become more tricky THEN.

Also a bug where the currently selected (in the same dialog) map save file
format was not being used when the current profile's map has been changed
has been fixed in this commit.

===========================================================================
To support the inclusion of other profile player location data in the map
file  MAP FILE FORMAT 18 (or greater) MUST BE CHOOSEN BEFORE THE MAP IS
COPIED.
===========================================================================

If this is not done then the map will not contain the data and then the
other profiles' player will be reset to the location of the player in the
currently active profile - as would happen for the old code (prior to
commit entitled "Enhance: store current player location in map file on a
per profile basis").  One bonus of this code is that only one copy of the
map to be copied over is made in the current profile's directory and all
the copies will have the same date-time stamp name between the profiles,
instead of a sequence of differently timed files as the file is copied one
profile at a time that code prior to this series of commits will produce!

Some items in the profile_preferences.ui have been renamed to clarify their
type or purpose or change thereof:

QGroupBox:
* groupBox_3               ==> groupBox_mapFiles

QLabel:
* label_9                  ==> label_saveMap
* label_10                 ==> label_loadMap
* label_12                 ==> label_copyMap
* map_file_action          ==> label_mapFileActionResult

QPushButtons:
* copy_map_profile         ==> pushButton_copyMap
* load_map_button          ==> pushButton_loadMap
* save_map_button          ==> pushButton_saveMap

Changes:
* (QComboBox) mapper_profiles_combobox
                           ==> (QPushButton) pushButton_chooseProfiles

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>