Tweak: make a QLabel widget span width of Profile Preferences dialog
Recent code changes uses longer messages in the "label_mapFileActionResult"
in the "Mapper" tab of the "Profile Preferences" dialog, it was however
only occupying one of the three columns in the QGridLayout in the part of
the dialogue where it was place - so that when a long message is displayed
it caused all the surrounding widgets in the other columns in the same
layout to be pushed out of the way.  This tweak makes it occupy ALL three
columns so this no longer occurs!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>